requirements.txt
Flask>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
certifi>=2024.0.0
DBUtils>=3.0.0
cachetools>=5.3.0
